python-mqtt-client/requirements.txt
paho-mqtt==1.5.1
pandas==1.2.0
matplotlib==3.3.3
tinydb==4.3.0
python_dotenv==0.15.0
